+++ Episodio.h
@@ -13,7 +13,7 @@ private:
 
 public:
     Episodio();
-    Episodio(string id, string nombre, string fecha_estreno, double cali, int duracion, int temporada, int numEp);
+    Episodio(string id, string nombre, string fecha_estreno, float cali, int duracion, int temporada, int numEp);
     int getTemp();
     int getNoEp();
     void display();
